Towards 2.03beta3
Inches/mm setting is now per-channel and saved by push/pop
When an unsupported G- or M-command is encountered, look for a macro of that name
Removed special support for M650 and M651, use macros instead
M203 always reports accelerations in mm/sec^2
Bug fix: G30 Pnn S-1 could report standard deviation as nan
Fixed direction for motors supported over CAN
Allow movement slightly outside M208 limits to allow for round-trip errors when the M208 limits are converted to motor positions and back again

+++ src/CAN/CanInterface.cpp
@@ -53,7 +53,7 @@ void CanInterface::StartMovement(const DDA& dda)
 
 // This is called by DDA::Prepare for each active CAN DM in the move
 // If steps == 0 then the drivers just need to be enabled
-void CanInterface::AddMovement(const DDA& dda, const PrepParams& params, size_t canDriver, size_t steps)
+void CanInterface::AddMovement(const DDA& dda, const PrepParams& params, size_t canDriver, int32_t steps)
 {
 	const size_t expansionBoardNumber = canDriver/DriversPerCanBoard;
 	CanMessageBuffer*& buf = movementBuffers[expansionBoardNumber];
